requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
pyautogui>=0.9.54
pyobjc-framework-Quartz>=9.2 